Move firebase query to its own file of queries from now on.
Separate Note type to its own file.

Add useEffect so every time a note is updated, it calls again the getNotes query, that produces when a note is deleted, it gets again the notes so you can see the update.
