+++ src/types/Note.ts
@@ -1,0 +1,7 @@
+type Note = {
+  title: string;
+  description: string;
+  create_timestamp: number;
+};
+
+export default Note;
